Normalize indentation
The code was originally indented by 20 spaces and tabified before the disassembly was published using Notepad++'s "Space to TAB" command (with tabstop=8), rendering it as 2 tabs and 4 following spaces.

This commit normalizes indentation to 16 spaces (equivalent to 2 tabs with tapstop=8), which is typical assembly style and cleanly divisible by 8.